# snapshot socket: improve performances