# snapshot socket: improve performances

Creates the BPF iterators only once for all network namespaces.

This patch improves the performances of the snapshot socket gadget by 98.93% when we have 100 containers.

## How to use

```
$ git checkout alban_perf_snapshot_socket
$ go test -exec sudo \
    -bench='^BenchmarkAllGadgetsWithContainers$/^container100$/snapshot-socket' \
    -run=Benchmark -count 8 \
    ./internal/benchmarks/...  | tee patched.bench
$ git checkout main
$ go test -exec sudo \
    -bench='^BenchmarkAllGadgetsWithContainers$/^container100$/snapshot-socket' \
    -run=Benchmark -count 8 \
    ./internal/benchmarks/...  | tee main.bench
$ benchstat main.bench patched.bench
goos: linux
goarch: amd64
pkg: github.com/inspektor-gadget/inspektor-gadget/internal/benchmarks
cpu: Intel(R) Core(TM) i7-6500U CPU @ 2.50GHz
                                                        │   main.bench   │           patched.bench            │
                                                        │     sec/op     │   sec/op     vs base               │
AllGadgetsWithContainers/container100/snapshot-socket-4   40972.4m ± 12%   439.1m ± 7%  -98.93% (p=0.000 n=8)
```

## Testing done

I tested manually with 2 containers that it reported tcp and udp 
```
$ go run -exec sudo ./cmd/ig/... snapshot socket --proto tcp
$ go run -exec sudo ./cmd/ig/... snapshot socket --proto udp
$ go run -exec sudo ./cmd/ig/... snapshot socket --proto all
```