@alban 2023-03-20
> Dividing times per 2 is clearly something I approve!

Dividing times by 93 :)