@alban 2023-03-21
Updated with @blanquicet 's patch. I tested that it still work for both `ig` and `kubectl-gadget`, and the benchmarks still look good.